// src/uWis/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export enum AssetStatus {
    NOT_FOUND = 'NotFound',
    UNKNOWN = 'Unknown',
    PREPARING = 'Preparing',
    ERRORED = 'Errored',
    READY = 'Ready'
}

export type GetStatusData = {
    body?: never;
    path?: never;
    query?: {
        assetId?: string;
    };
    url: '/umbraco/uwis/api/v1/status';
};

export type GetStatusErrors = {
    /**
     * The resource is protected and requires an authentication token
     */
    401: unknown;
};

export type GetStatusResponses = {
    /**
     * OK
     */
    200: AssetStatus;
};

export type GetStatusResponse = GetStatusResponses[keyof GetStatusResponses];

export type ClientOptions = {
    baseUrl: 'http://localhost:59927' | (string & {});
};